+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 
 **Changed**
 
-- [`HtmlToken#toHtml`](https://github.com/bhsd-harry/wikiparser-node/wiki/FileToken-%28EN%29#tohtml) now renders `<h1>` to `<h6>` tags with a `mw-html-heading` class
+- [`HtmlToken#toHtml`](https://github.com/bhsd-harry/wikiparser-node/wiki/HtmlToken-%28EN%29#tohtml) now renders `<h1>` to `<h6>` tags with a `mw-html-heading` class
+- [`FileToken#toHtml`](https://github.com/bhsd-harry/wikiparser-node/wiki/FileToken-%28EN%29#tohtml) now adds `loading="lazy"` to rendered `<img>` tags
 
 ## v1.46.3